mqtt_transport: Ping every 8 seconds instead of 15
If there's a large interval between messages being transmitted, it
appears that the 15 second interval is not sufficient to keep the
connection alive. Testing the temperature-client with 10 second interval
seems to work (and 90 second intervals for both data streams).

Set the interval to 8 to leave some room for network slowness.